Remove unnecessary methods and use the filter on the messages instead
